+++ libs/table/features/src/index.ts
@@ -5,3 +5,4 @@ export * from './table-selectable-feature';
 export * from './table-total-feature';
 export * from './table-pagination-feature';
 export * from './table-row-actions-feature';
+export * from './table-batch-actions-feature';