+++ libs/button/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-next.0](https://github.com/spryker/ui-components/compare/@spryker/button@0.2.0-fix-web-components.0...@spryker/button@0.2.0-next.0) (2020-07-21)
+
+
+### Bug Fixes
+
+* **libs:** update versions for next release ([9998cae](https://github.com/spryker/ui-components/commit/9998cae9b2ab631607c0d33fa546363313bfd6aa))
+
+
+
+
+
+# [0.1.0-fix-web-components.2](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-fix-web-components.1...@spryker/button@0.1.0-fix-web-components.2) (2020-07-20)
+
+
+### Bug Fixes
+
+* **button:** allow buttons to be sized from host element ([d6a53d6](https://github.com/spryker/ui-components/commit/d6a53d60c3aa9c134cd53a80b588b9eb52b7d774))
+
+
+
+
+
+# [0.1.0-fix-web-components.1](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-fix-web-components.0...@spryker/button@0.1.0-fix-web-components.1) (2020-07-16)
+
+
+### Bug Fixes
+
+* **lib:** update all fixed package peer versions ([bebf900](https://github.com/spryker/ui-components/commit/bebf900c4867617f4dd0032a554037827ecdbda6))
+
+
+
+
+
+# [0.1.0-fix-web-components.0](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.17...@spryker/button@0.1.0-fix-web-components.0) (2020-07-16)
+
+
+### Bug Fixes
+
+* **lib:** update fixed libs versions ([7a1fc0a](https://github.com/spryker/ui-components/commit/7a1fc0aaf1b949ef0886f7e2b8c54591f4867a77))
+
+
+
+
+
+# [0.1.0-next.17](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.16...@spryker/button@0.1.0-next.17) (2020-07-15)
+
+**Note:** Version bump only for package @spryker/button
+
+
+
+
+
+# [0.1.0-next.16](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.15...@spryker/button@0.1.0-next.16) (2020-07-14)
+
+
+### Features
+
+* **button:** added import for new component ([66aa501](https://github.com/spryker/ui-components/commit/66aa501e698cd14d8541c57198c865b98941e2c9))
+* **button:** code refactoring ([b8aa3b9](https://github.com/spryker/ui-components/commit/b8aa3b93e62db6da02da1218765d465536da7c4f))
+* **button:** code refactoring ([5bdead0](https://github.com/spryker/ui-components/commit/5bdead04d287292797a60005c9845f54f85a2716))
+
+
+
+
+
 # [0.1.0-next.15](https://github.com/spryker/ui-components/compare/@spryker/button@0.1.0-next.14...@spryker/button@0.1.0-next.15) (2020-07-11)